LPD-71504 Make testSmoke() run for all DL store implementations
