+++ modules/apps/portal-store/portal-store-s3-test/src/testIntegration/java/com/liferay/portal/store/s3/test/IBMS3StoreTest.java
@@ -6,25 +6,16 @@
 package com.liferay.portal.store.s3.test;
 
 import com.liferay.arquillian.extension.junit.bridge.junit.Arquillian;
-import com.liferay.document.library.kernel.model.DLFileEntry;
-import com.liferay.document.library.kernel.model.DLFileVersion;
-import com.liferay.document.library.kernel.service.DLTrashLocalServiceUtil;
 import com.liferay.document.library.kernel.store.Store;
-import com.liferay.document.library.test.util.DLAppTestUtil;
 import com.liferay.petra.io.DummyOutputStream;
 import com.liferay.petra.io.StreamUtil;
 import com.liferay.petra.io.unsync.UnsyncByteArrayInputStream;
 import com.liferay.petra.string.StringBundler;
 import com.liferay.petra.string.StringPool;
-import com.liferay.portal.kernel.model.Group;
-import com.liferay.portal.kernel.repository.model.FileEntry;
 import com.liferay.portal.kernel.test.TestInfo;
 import com.liferay.portal.kernel.test.rule.AggregateTestRule;
 import com.liferay.portal.kernel.test.rule.AssumeTestRule;
-import com.liferay.portal.kernel.test.util.GroupTestUtil;
 import com.liferay.portal.kernel.test.util.RandomTestUtil;
-import com.liferay.portal.kernel.trash.TrashHandler;
-import com.liferay.portal.kernel.trash.TrashHandlerRegistryUtil;
 import com.liferay.portal.kernel.util.GetterUtil;
 import com.liferay.portal.kernel.util.PropsKeys;
 import com.liferay.portal.kernel.util.PropsUtil;
@@ -109,43 +100,6 @@ public class IBMS3StoreTest extends BaseStoreTestCase {
 				companyId, repositoryId, fileName, Store.VERSION_DEFAULT));
 
 		_store.deleteDirectory(companyId, repositoryId, StringPool.SLASH);
-	}
-
-	@Test
-	public void testSmoke() throws Exception {
-		Group group = GroupTestUtil.addGroup();
-
-		FileEntry fileEntry = DLAppTestUtil.addFileEntry(group.getGroupId());
-
-		DLFileEntry dlFileEntry = (DLFileEntry)fileEntry.getModel();
-
-		DLFileVersion dlFileVersion = dlFileEntry.getFileVersion();
-
-		Assert.assertTrue(
-			_store.hasFile(
-				dlFileEntry.getCompanyId(), dlFileEntry.getDataRepositoryId(),
-				dlFileEntry.getName(), dlFileVersion.getStoreFileName()));
-
-		DLTrashLocalServiceUtil.moveFileEntryToTrash(
-			dlFileEntry.getUserId(), dlFileEntry.getRepositoryId(),
-			dlFileEntry.getFileEntryId());
-
-		Assert.assertTrue(
-			_store.hasFile(
-				dlFileEntry.getCompanyId(), dlFileEntry.getDataRepositoryId(),
-				dlFileEntry.getName(), dlFileVersion.getStoreFileName()));
-
-		TrashHandler trashHandler = TrashHandlerRegistryUtil.getTrashHandler(
-			DLFileEntry.class.getName());
-
-		trashHandler.deleteTrashEntry(dlFileEntry.getPrimaryKey());
-
-		Assert.assertFalse(
-			_store.hasFile(
-				dlFileEntry.getCompanyId(), dlFileEntry.getDataRepositoryId(),
-				dlFileEntry.getName(), dlFileVersion.getStoreFileName()));
-
-		GroupTestUtil.deleteGroup(group);
 	}
 
 	@Override
